[Refactor/Layout] mainpage conponent 추가
