+++ src/app/layout.tsx
@@ -1,7 +1,7 @@
 import './globals.css';
-import Header from '@/components/header/Header';
+import Header from '@/components/layout/header/Header';
 import { Providers } from './providers';
-import Footer from '@/components/footer/Footer';
+import Footer from '@/components/layout/footer/Footer';
 
 export default function RootLayout({
   children,
