+++ src/components/mainpage/MainPage.tsx
@@ -1,0 +1,56 @@
+'use client';
+
+import styled from 'styled-components';
+
+const MainPageBox = styled.div`
+  margin-top: 108px;
+  @media screen and (max-width: 707px) {
+    margin-top: 222px;
+  }
+`;
+
+export default function MainPage() {
+  return (
+    <MainPageBox>
+      <h1>Hello World</h1>
+      <h1>Hello World</h1>
+      <h1>Hello World</h1>
+      <h1>Hello World</h1>
+      <h1>Hello World</h1>
+      <h1>Hello World</h1>
+      <h1>Hello World</h1>
+      <h1>Hello World</h1>
+      <h1>Hello World</h1>
+      <h1>Hello World</h1>
+      <h1>Hello World</h1>
+      <h1>Hello World</h1>
+      <h1>Hello World</h1>
+      <h1>Hello World</h1>
+      <h1>Hello World</h1>
+      <h1>Hello World</h1>
+      <h1>Hello World</h1>
+      <h1>Hello World</h1>
+      <h1>Hello World</h1>
+      <h1>Hello World</h1>
+      <h1>Hello World</h1>
+      <h1>Hello World</h1>
+      <h1>Hello World</h1>
+      <h1>Hello World</h1>
+      <h1>Hello World</h1>
+      <h1>Hello World</h1>
+      <h1>Hello World</h1>
+      <h1>Hello World</h1>
+      <h1>Hello World</h1>
+      <h1>Hello World</h1>
+      <h1>Hello World</h1>
+      <h1>Hello World</h1>
+      <h1>Hello World</h1>
+      <h1>Hello World</h1>
+      <h1>Hello World</h1>
+      <h1>Hello World</h1>
+      <h1>Hello World</h1>
+      <h1>Hello World</h1>
+      <h1>Hello World</h1>
+    </MainPageBox>
+  );
+}
